+++ lib/souls/generate/query.rb
@@ -25,6 +25,7 @@ module Souls
 
       def create_query class_name: "souls"
         file_path = "./app/graphql/queries/#{class_name}.rb"
+        return "Query already exist! #{file_path}" if File.exist? file_path
         File.open(file_path, "w") do |f|
           f.write <<~EOS
             module Queries
